long-term-storage/requirements.txt
# Requirements for pigasus, update as needed....
pytest
pylint
pytest-cov
requests
pymssql
python-dotenv
pandas
pyarrow
boto3